chore(#1051): delete vibew deploy orchestration
Remove the remote SSH orchestration surface sunset in ADR-086.

Deleted files (~4 900 LOC):
- internal/cli/cmd/deploy.go + deploy_test.go + deploy_project_name_test.go
- internal/app/deploy/{service_test,multiapp,detect,health,image_transfer,
  openbao,arch,errors}.go and their tests

service.go is reduced to the bare Service shell (struct + NewService +
WithBundleFS + WithImageSaver) plus ProjectNameFromConfig, which bundle.go
still depends on. A new helpers.go preserves SidecarComposeData,
AppComposeData, appContainerName, isLocalUpstreamHost, and
defaultHealthPort — all still consumed by bundle.go.

artifact_test.go is filtered to keep only the Bundle-focused cases; the
orchestration cases that drove svc.Deploy / BootstrapSidecar /
DeployMultiApp are dropped. testutil_test.go provides the minimal
fakeExecutor / fakeGenerator doubles previously defined in
service_test.go.

Stub cobra command and package rename follow in the next two commits.
