+++ internal/app/deploy/service.go
@@ -1,182 +1,44 @@
-// Package deploy provides the application service that deploys a VibeWarden
-// project to a remote server over SSH.
+// Package deploy provides the application service that produces the
+// `vibew bundle` deployment artifact.
+//
+// The package name is a historical artefact from the removed `vibew deploy`
+// command and is renamed to `bundle` in a follow-up commit on this branch
+// (ADR-086).
 package deploy
 
 import (
-	"context"
-	"errors"
-	"fmt"
-	"io"
-	"io/fs"
 	"os"
 	"path/filepath"
 	"strings"
-	"time"
 
-	"github.com/vibewarden/vibewarden/internal/config"
 	"github.com/vibewarden/vibewarden/internal/ports"
 )
 
-// ErrHealthCheck is returned when the sidecar health check fails after
-// deployment. Files have been deployed and services started, but the
-// sidecar did not become healthy within the timeout.
-var ErrHealthCheck = errors.New("health check failed")
-
-// DriftError is returned when remote files have diverged from local files
-// and --force was not specified. It carries the list of changes that would
-// be applied so the CLI can display them to the user. Changes are categorised
-// into deploy-owned files (generated by the bundle and safe to overwrite) and
-// user files (potentially hand-edited on the server).
-type DriftError struct {
-	// Changes is a list of human-readable descriptions of files that would
-	// be modified or deleted on the remote host.
-	Changes []string
-}
-
-// deployOwnedFiles is the set of file paths (relative to the deploy root)
-// that are generated by the deploy bundle. Changes to these files are expected
-// on redeploy and are categorised separately from user-modified files.
-var deployOwnedFiles = map[string]bool{
-	"docker-compose.yml": true,
-	"vibewarden.yaml":    true,
-	".credentials":       true,
-	".env":               true,
-}
-
-// deployOwnedPrefixes is the set of directory prefixes that indicate a file
-// belongs to the deploy bundle (e.g. "kratos/", "openbao/").
-var deployOwnedPrefixes = []string{
-	"kratos/",
-	"openbao/",
-}
-
-// isDeployOwned returns true when the filename (extracted from an rsync
-// itemize-changes line) corresponds to a deploy-generated file.
-func isDeployOwned(filename string) bool {
-	if deployOwnedFiles[filename] {
-		return true
-	}
-	for _, prefix := range deployOwnedPrefixes {
-		if strings.HasPrefix(filename, prefix) {
-			return true
-		}
-	}
-	return false
-}
-
-// extractFilename returns the filename portion of an rsync --itemize-changes
-// line. The format is "<code> <filename>", where the code is the first
-// space-delimited field.
-func extractFilename(line string) string {
-	if idx := strings.IndexByte(line, ' '); idx != -1 {
-		return strings.TrimSpace(line[idx+1:])
-	}
-	return line
-}
-
-// Error implements the error interface. It categorises changes into
-// deploy-owned files and user-modified files for clearer diagnostics.
-func (e *DriftError) Error() string {
-	var deployOwned, userModified []string
-	for _, c := range e.Changes {
-		filename := extractFilename(c)
-		if isDeployOwned(filename) {
-			deployOwned = append(deployOwned, c)
-		} else {
-			userModified = append(userModified, c)
-		}
-	}
-
-	var b strings.Builder
-	b.WriteString("remote files have been modified since last deploy:\n")
-
-	if len(userModified) > 0 {
-		b.WriteString("\n  User-modified files (may contain manual changes):\n")
-		for _, c := range userModified {
-			b.WriteString("    ")
-			b.WriteString(c)
-			b.WriteString("\n")
-		}
-	}
-
-	if len(deployOwned) > 0 {
-		b.WriteString("\n  Deploy-owned files (safe to overwrite):\n")
-		for _, c := range deployOwned {
-			b.WriteString("    ")
-			b.WriteString(c)
-			b.WriteString("\n")
-		}
-	}
-
-	b.WriteString("\nRun 'vibew deploy --force' to overwrite, or back up changes first.")
-	return b.String()
-}
-
-const (
-	// remoteBaseDir is the root directory on the remote host where all
-	// VibeWarden projects are deployed.
-	remoteBaseDir = "~/vibewarden"
-
-	// healthCheckTimeout is the maximum time to wait for the sidecar to become
-	// healthy after starting Docker Compose.
-	healthCheckTimeout = 60 * time.Second
-
-	// healthCheckInterval is the delay between successive health check attempts.
-	healthCheckInterval = 3 * time.Second
-
-	// defaultHealthPort is the port used when cfg.Server.Port is zero.
-	defaultHealthPort = 8443
-)
-
-// Service orchestrates the "vibew deploy" use case.
-// It generates runtime config, transfers files to the remote, starts Docker
-// Compose, and verifies the sidecar health endpoint.
+// Service produces the `vibew bundle` deployment artifact.
 type Service struct {
-	executor      ports.RemoteExecutor
-	generator     ports.ConfigGenerator
-	imageExporter ports.ImageExporter
+	generator ports.ConfigGenerator
 
-	// bundleFS is the filesystem adapter used by the "vibew bundle" extras
-	// pipeline (sample.env, .env, deploy.sh, README.md). When nil, the
-	// extras pipeline falls back to the real OS filesystem via the
-	// default writeFile / os.Stat helpers so existing callers of
-	// Service.Bundle (notably vibew deploy --dry-run) keep working unchanged.
+	// bundleFS is the filesystem adapter used by the bundle extras pipeline
+	// (sample.env, .env, deploy.sh, README.md). When nil, the extras pipeline
+	// is a no-op.
 	bundleFS ports.BundleFS
 
-	// imageSaver saves the app image to image.tar during vibew bundle.
-	// When nil, the extras pipeline treats --skip-image as always-on for
-	// that single invocation, which matches the behaviour requested by
-	// callers that do not wire an ImageSaver (e.g. vibew deploy --dry-run).
+	// imageSaver saves the app image to image.tar during vibew bundle. When
+	// nil, the extras pipeline treats --skip-image as always-on for the
+	// invocation.
 	imageSaver ports.ImageSaver
-
-	// localArch overrides runtime.GOARCH for testing. When empty, the real
-	// runtime.GOARCH is used. Set via WithLocalArch.
-	localArch string
 }
 
 // NewService creates a Service.
-// executor handles SSH commands and rsync transfers.
-// generator is used to produce the .vibewarden/generated/ files before transfer.
-func NewService(
-	executor ports.RemoteExecutor,
-	generator ports.ConfigGenerator,
-) *Service {
-	return &Service{
-		executor:  executor,
-		generator: generator,
-	}
+//
+// generator is used to produce the generated/ files referenced by the
+// bundle. executor is accepted for backwards compatibility with existing
+// call sites that pass a nil value; the bundle pipeline does not use it.
+func NewService(_ ports.RemoteExecutor, generator ports.ConfigGenerator) *Service {
+	return &Service{generator: generator}
 }
 
-// WithImageExporter sets the ImageExporter on the Service and returns it for
-// chaining. The exporter is used to save Docker images from the local daemon so
-// they can be transferred to the remote host when the image name has no registry
-// prefix (bare name like "myapp:latest").
-func (s *Service) WithImageExporter(exporter ports.ImageExporter) *Service {
-	s.imageExporter = exporter
-	return s
-}
-
-// WithBundleFS sets the filesystem adapter used by the vibew bundle extras
+// WithBundleFS sets the filesystem adapter used by the bundle extras
 // pipeline and returns the Service for chaining. Production callers pass a
 // bundlefs.OSFS; tests pass an in-memory fake.
 func (s *Service) WithBundleFS(bfs ports.BundleFS) *Service {
@@ -184,523 +46,21 @@ func (s *Service) WithBundleFS(bfs ports.BundleFS) *Service {
 	return s
 }
 
-// WithImageSaver sets the ImageSaver used by the vibew bundle extras
-// pipeline and returns the Service for chaining. When not set, image.tar is
-// never produced (equivalent to --skip-image). Production callers pass an
-// opsadapter.ImageExportAdapter; tests pass a fake.
+// WithImageSaver sets the ImageSaver used by the bundle extras pipeline and
+// returns the Service for chaining. When not set, image.tar is never
+// produced (equivalent to --skip-image).
 func (s *Service) WithImageSaver(saver ports.ImageSaver) *Service {
 	s.imageSaver = saver
 	return s
 }
 
-// WithLocalArch overrides the local architecture used for arch mismatch
-// detection. This is intended for testing — in production the real
-// runtime.GOARCH is used. Returns the Service for chaining.
-func (s *Service) WithLocalArch(arch string) *Service {
-	s.localArch = arch
-	return s
-}
-
-// RunOptions holds parameters for a deploy run.
-type RunOptions struct {
-	// ConfigPath is the path to vibewarden.yaml on the local filesystem.
-	ConfigPath string
-
-	// ProdConfigPath is the optional path to the production override file
-	// (e.g. vibewarden.production.yaml). When set, its values are deep-merged
-	// on top of the base config before writing to the deploy bundle.
-	ProdConfigPath string
-
-	// ProjectName is used as the remote sub-directory name under remoteBaseDir.
-	// When empty it is derived from the basename of the directory containing
-	// ConfigPath.
-	ProjectName string
-
-	// GeneratedDir is the local directory where generated files are written
-	// before transfer. Defaults to ".vibewarden/generated" when empty.
-	GeneratedDir string
-
-	// Env is the deployment environment name (e.g. "production", "staging").
-	// Defaults to "production" when empty.
-	Env string
-
-	// Force, when true, skips the drift detection warning and overwrites
-	// remote files unconditionally. When false (the default), a dry-run rsync
-	// is performed before the real transfer and the deploy aborts with an
-	// actionable message if remote files have diverged.
-	Force bool
-
-	// Out is the writer used for progress messages. May be nil (output is
-	// discarded).
-	Out io.Writer
-}
-
-// Deploy runs the full deployment flow:
-//  1. Bundle all deploy files locally into .vibewarden/deploy/
-//  2. Verify Docker + Docker Compose on the remote
-//  3. rsync .vibewarden/deploy/ to the remote
-//     When app.build is set, the app source directory is also transferred so
-//     that Docker can build the image remotely.
-//  4. docker compose up -d --build  (build mode) or
-//     docker compose pull && docker compose up -d  (image mode)
-//  5. Health check
-func (s *Service) Deploy(ctx context.Context, cfg *config.Config, opts RunOptions) error {
-	out := opts.Out
-	if out == nil {
-		out = io.Discard
-	}
-
-	projectName := opts.ProjectName
-	if projectName == "" {
-		projectName = ProjectNameFromConfig(opts.ConfigPath)
-	}
-	remoteDir := remoteBaseDir + "/" + projectName + "/"
-
-	// Step 1: produce the local deploy bundle.
-	fmt.Fprintln(out, "Generating deploy bundle...")
-	bundleDir := opts.GeneratedDir
-	if bundleDir == "" {
-		bundleDir = defaultBundleDir
-	}
-	if err := s.Bundle(ctx, BundleOptions{
-		Config:         cfg,
-		ConfigPath:     opts.ConfigPath,
-		ProdConfigPath: opts.ProdConfigPath,
-		ProjectName:    projectName,
-		MultiSite:      false,
-		OutputDir:      bundleDir,
-		Env:            opts.Env,
-	}); err != nil {
-		return fmt.Errorf("creating deploy bundle: %w", err)
-	}
-
-	// Determine whether a locally-built image will be transferred to the
-	// remote. This is needed for the arch mismatch check in prerequisites.
-	imageName := cfg.App.Image
-	if cfg.App.Build != "" {
-		imageName = cfg.ComposeProjectName() + "-app:latest"
-	}
-	willTransferLocalImage := imageName != "" && isLocalImage(imageName) && s.imageExporter != nil
-
-	// Step 2: verify prerequisites on the remote.
-	fmt.Fprintln(out, "Verifying remote prerequisites...")
-	if err := s.checkRemotePrerequisites(ctx, willTransferLocalImage); err != nil {
-		return fmt.Errorf("remote prerequisites check failed: %w", err)
-	}
-
-	// Step 3: transfer files.
-	fmt.Fprintf(out, "Transferring files to remote %s...\n", remoteDir)
-
-	// Ensure the remote directory exists.
-	if _, err := s.executor.Run(ctx, "mkdir -p "+remoteDir); err != nil {
-		return fmt.Errorf("creating remote directory: %w", err)
-	}
-
-	// On redeploy, preserve existing credentials. The bundle generates fresh
-	// random passwords, but the DB volume retains the originals. Downloading
-	// the remote credential files into the bundle BEFORE drift detection
-	// ensures the dry-run compares identical credential content and does not
-	// flag these files as false-positive drift. (#991, #1031)
-	for _, credFile := range []string{".credentials", ".env"} {
-		remotePath := remoteDir + credFile
-		localPath := filepath.Join(bundleDir, credFile)
-		if _, fetchErr := s.executor.Run(ctx, "test -f "+remotePath); fetchErr == nil {
-			credData, _ := s.executor.Run(ctx, "cat "+remotePath)
-			if credData != "" {
-				_ = os.WriteFile(localPath, []byte(credData), 0o600) //nolint:gosec
-			}
-		}
-	}
-
-	// Drift detection: before overwriting remote files with --delete, check
-	// what would change and abort unless --force is set. Credential
-	// preservation above ensures deploy-owned credential files are not
-	// reported as drift.
-	if !opts.Force {
-		changes, err := s.executor.DryRunTransfer(ctx, bundleDir, remoteDir)
-		if err != nil {
-			// If the dry-run fails (e.g. remote directory does not exist yet on
-			// first deploy), fall through -- the real rsync will create it.
-			fmt.Fprintf(out, "Note: drift detection skipped (%v)\n", err)
-		} else if len(changes) > 0 {
-			return &DriftError{Changes: changes}
-		}
-	}
-
-	// rsync the deploy bundle.
-	if err := s.executor.Transfer(ctx, bundleDir, remoteDir, true); err != nil {
-		return fmt.Errorf("transferring deploy bundle: %w", err)
-	}
-
-	// Step 4: transfer local image.
-	// When app.build is set the image was built locally by `vibew build`.
-	// Derive the image name (projectName-app:latest) and transfer it via
-	// docker save/load. No source code is transferred to the server.
-	localImageTransferred := false
-	if willTransferLocalImage {
-		if err := s.transferLocalImage(ctx, imageName, remoteDir, out); err != nil {
-			return fmt.Errorf("transferring local image: %w", err)
-		}
-		localImageTransferred = true
-	}
-
-	// Step 5: start Docker Compose on the remote.
-	// Always pull the latest sidecar image before starting, regardless of mode.
-	fmt.Fprintln(out, "Pulling latest sidecar image on remote...")
-	pullSidecarCmd := fmt.Sprintf("cd %s && docker compose pull vibewarden", remoteDir)
-	if _, err := s.executor.Run(ctx, pullSidecarCmd); err != nil {
-		return fmt.Errorf("docker compose pull vibewarden: %w", err)
-	}
-
-	if localImageTransferred {
-		// Local image was already transferred -- skip pulling and just start.
-		fmt.Fprintln(out, "Starting services on remote...")
-		upCmd := fmt.Sprintf("cd %s && docker compose up -d --force-recreate", remoteDir)
-		if _, err := s.executor.Run(ctx, upCmd); err != nil {
-			return fmt.Errorf("docker compose up: %w", err)
-		}
-	} else {
-		fmt.Fprintln(out, "Pulling Docker images on remote...")
-		pullCmd := fmt.Sprintf("cd %s && docker compose pull", remoteDir)
-		if _, err := s.executor.Run(ctx, pullCmd); err != nil {
-			return fmt.Errorf("docker compose pull: %w", err)
-		}
-
-		fmt.Fprintln(out, "Starting services on remote...")
-		upCmd := fmt.Sprintf("cd %s && docker compose up -d --force-recreate", remoteDir)
-		if _, err := s.executor.Run(ctx, upCmd); err != nil {
-			return fmt.Errorf("docker compose up: %w", err)
-		}
-	}
-
-	// Step 6: health check -- run curl on the remote so the probe is independent
-	// of DNS propagation, external port availability, and TLS certificate issuance.
-	// Use the merged config (base + prod overlay) for the correct port and TLS state.
-	// When opts.ConfigPath is empty (programmatic callers that build cfg in-memory)
-	// or when the file is not on disk (unit tests), fall back to the supplied cfg.
-	healthCfg := cfg
-	if opts.ConfigPath != "" {
-		_, statErr := os.Stat(opts.ConfigPath)
-		switch {
-		case statErr == nil:
-			loaded, err := LoadMergedConfig(opts.ConfigPath, opts.ProdConfigPath)
-			if err != nil {
-				return err
-			}
-			healthCfg = loaded
-		case errors.Is(statErr, fs.ErrNotExist):
-			// Intentional fall-through: caller supplied an in-memory cfg
-			// (programmatic or unit-test entrypoint); no disk file to merge.
-		default:
-			return fmt.Errorf("stat config %s: %w", opts.ConfigPath, statErr)
-		}
-	}
-	port := healthCfg.Server.Port
-	if port == 0 {
-		port = defaultHealthPort
-	}
-	healthURL := healthCheckURL(port, healthCfg.TLS.Enabled)
-	fmt.Fprintf(out, "Waiting for sidecar health check at %s (via SSH)...\n", healthURL)
-	if !s.waitHealthy(ctx, port, healthCfg.TLS.Enabled, out) {
-		fmt.Fprintln(out, "Deploy completed but health check failed — running diagnostics...")
-		diagnostic := s.diagnoseHealthFailure(ctx, remoteDir, port, healthCfg.TLS.Enabled)
-		fmt.Fprint(out, "\n")
-		fmt.Fprint(out, FormatDiagnostic(diagnostic))
-		return &HealthCheckError{Diagnostic: diagnostic}
-	}
-
-	fmt.Fprintln(out, "Deploy complete.")
-	return nil
-}
-
-// StatusOptions holds parameters for the deploy status command.
-type StatusOptions struct {
-	// ConfigPath is the path to vibewarden.yaml on the local filesystem.
-	// It is used to derive the project name (and therefore the remote directory)
-	// in exactly the same way as RunOptions.ConfigPath in Deploy. When empty,
-	// the project name falls back to "vibewarden".
-	ConfigPath string
-
-	// ProjectName overrides the project name derived from ConfigPath.
-	ProjectName string
-
-	// Out is the writer used for status output. May be nil.
-	Out io.Writer
-}
-
-// Status fetches Docker Compose service state from the remote.
-func (s *Service) Status(ctx context.Context, opts StatusOptions) error {
-	out := opts.Out
-	if out == nil {
-		out = io.Discard
-	}
-
-	projectName := opts.ProjectName
-	if projectName == "" {
-		projectName = ProjectNameFromConfig(opts.ConfigPath)
-	}
-	remoteDir := remoteBaseDir + "/" + projectName + "/"
-
-	output, err := s.executor.Run(ctx, "docker compose --project-directory "+remoteDir+" ps")
-	if err != nil {
-		return fmt.Errorf("fetching remote status: %w", err)
-	}
-	fmt.Fprintln(out, output)
-	return nil
-}
-
-// LogsOptions holds parameters for the deploy logs command.
-type LogsOptions struct {
-	// ConfigPath is the path to vibewarden.yaml on the local filesystem.
-	// It is used to derive the project name (and therefore the remote directory)
-	// in exactly the same way as RunOptions.ConfigPath in Deploy. When empty,
-	// the project name falls back to "vibewarden".
-	ConfigPath string
-
-	// ProjectName overrides the project name derived from ConfigPath.
-	ProjectName string
-
-	// Lines is the number of log lines to retrieve (0 = all).
-	Lines int
-
-	// Follow streams new log lines continuously, like "docker compose logs -f".
-	// When true the command runs until the context is cancelled (e.g. Ctrl-C).
-	// Output is written directly to Out in real-time without buffering.
-	Follow bool
-
-	// Out is the writer used for log output. May be nil.
-	Out io.Writer
-}
-
-// Logs retrieves Docker Compose logs from the remote.
-// When opts.Follow is true the command streams log output in real-time by
-// using RunStream; the call blocks until the context is cancelled. When false
-// the output is fetched in a single buffered Run call and written to opts.Out.
-func (s *Service) Logs(ctx context.Context, opts LogsOptions) error {
-	out := opts.Out
-	if out == nil {
-		out = io.Discard
-	}
-
-	projectName := opts.ProjectName
-	if projectName == "" {
-		projectName = ProjectNameFromConfig(opts.ConfigPath)
-	}
-	remoteDir := remoteBaseDir + "/" + projectName + "/"
-
-	cmd := "docker compose --project-directory " + remoteDir + " logs"
-	if opts.Lines > 0 {
-		cmd += fmt.Sprintf(" --tail=%d", opts.Lines)
-	}
-	if opts.Follow {
-		cmd += " -f"
-		if err := s.executor.RunStream(ctx, cmd, out, out); err != nil {
-			return fmt.Errorf("streaming remote logs: %w", err)
-		}
-		return nil
-	}
-
-	output, err := s.executor.Run(ctx, cmd)
-	if err != nil {
-		return fmt.Errorf("fetching remote logs: %w", err)
-	}
-	fmt.Fprintln(out, output)
-	return nil
-}
-
-// ListSites returns the names of all site subdirectories under
-// ~/vibewarden/sites/ on the remote host. An empty slice is returned
-// when no sites exist or the directory is missing.
-func (s *Service) ListSites(ctx context.Context) ([]string, error) {
-	output, err := s.executor.Run(ctx, "ls -1 "+sitesDir+" 2>/dev/null || true")
-	if err != nil {
-		return nil, fmt.Errorf("listing remote sites: %w", err)
-	}
-	output = strings.TrimSpace(output)
-	if output == "" {
-		return nil, nil
-	}
-	return strings.Split(output, "\n"), nil
-}
-
-// StatusMultiApp fetches Docker Compose service state for multi-app deployments.
-// When app is non-empty, it targets that specific site. When app is empty, it
-// shows the sidecar status plus all site statuses.
-func (s *Service) StatusMultiApp(ctx context.Context, app string, out io.Writer) error {
-	if out == nil {
-		out = io.Discard
-	}
-
-	if app != "" {
-		siteDir := sitesDir + app + "/"
-		output, err := s.executor.Run(ctx, "docker compose --project-directory "+siteDir+" ps")
-		if err != nil {
-			return fmt.Errorf("fetching status for site %q: %w", app, err)
-		}
-		fmt.Fprintf(out, "=== Site: %s ===\n", app)
-		fmt.Fprintln(out, output)
-		return nil
-	}
-
-	// Show sidecar status.
-	sidecarOutput, err := s.executor.Run(ctx, "docker compose --project-directory "+sidecarDir+" ps")
-	if err != nil {
-		return fmt.Errorf("fetching sidecar status: %w", err)
-	}
-	fmt.Fprintln(out, "=== Sidecar ===")
-	fmt.Fprintln(out, sidecarOutput)
-
-	// Enumerate and show each site.
-	sites, err := s.ListSites(ctx)
-	if err != nil {
-		return err
-	}
-	for _, name := range sites {
-		siteDir := sitesDir + name + "/"
-		output, siteErr := s.executor.Run(ctx, "docker compose --project-directory "+siteDir+" ps")
-		fmt.Fprintf(out, "=== Site: %s ===\n", name)
-		if siteErr != nil {
-			fmt.Fprintf(out, "  error: %v\n", siteErr)
-			continue
-		}
-		fmt.Fprintln(out, output)
-	}
-	return nil
-}
-
-// LogsMultiApp fetches Docker Compose logs for multi-app deployments.
-// When app is non-empty, it targets that specific site. When app is empty, it
-// shows logs from the sidecar container.
-func (s *Service) LogsMultiApp(ctx context.Context, app string, opts LogsOptions) error {
-	out := opts.Out
-	if out == nil {
-		out = io.Discard
-	}
-
-	var dir string
-	if app != "" {
-		dir = sitesDir + app + "/"
-	} else {
-		dir = sidecarDir
-	}
-
-	cmd := "docker compose --project-directory " + dir + " logs"
-	if opts.Lines > 0 {
-		cmd += fmt.Sprintf(" --tail=%d", opts.Lines)
-	}
-	if opts.Follow {
-		cmd += " -f"
-		if err := s.executor.RunStream(ctx, cmd, out, out); err != nil {
-			return fmt.Errorf("streaming remote logs: %w", err)
-		}
-		return nil
-	}
-
-	output, err := s.executor.Run(ctx, cmd)
-	if err != nil {
-		return fmt.Errorf("fetching remote logs: %w", err)
-	}
-	fmt.Fprintln(out, output)
-	return nil
-}
-
-// checkRemotePrerequisites verifies that docker and docker compose are available
-// on the remote host. When willTransferLocalImage is true, it additionally
-// checks that the local build architecture matches the remote server
-// architecture, returning an ArchMismatchError with a fix-it suggestion if they
-// differ.
-func (s *Service) checkRemotePrerequisites(ctx context.Context, willTransferLocalImage bool) error {
-	if _, err := s.executor.Run(ctx, "which docker"); err != nil {
-		return fmt.Errorf("docker not found on remote: %w", err)
-	}
-	if _, err := s.executor.Run(ctx, "docker compose version"); err != nil {
-		return fmt.Errorf("docker compose not found on remote: %w", err)
-	}
-	if willTransferLocalImage {
-		if err := s.checkArchCompatibility(ctx); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// waitHealthy polls the sidecar health endpoint on the remote host via SSH
-// until curl reports success (exit code 0) or the context deadline /
-// healthCheckTimeout expires.
-//
-// When tlsEnabled is true the probe uses HTTPS with the -k flag (skip TLS
-// certificate verification) because we are hitting localhost, not the public
-// domain. When TLS is disabled the probe uses plain HTTP.
-//
-// Running the probe over SSH avoids all DNS propagation, firewall, and TLS
-// certificate issuance dependencies — the request is made from inside the
-// server to its own localhost interface.
-//
-// Returns true when the sidecar becomes healthy, false on timeout or context
-// cancellation. When false, a warning is printed suggesting diagnostic
-// commands the operator can run.
-func (s *Service) waitHealthy(ctx context.Context, port int, tlsEnabled bool, out io.Writer) bool {
-	cmd := healthCheckCmd(port, tlsEnabled)
-	deadline := time.Now().Add(healthCheckTimeout)
-	attempt := 0
-	var lastErr error
-
-	for {
-		attempt++
-		_, err := s.executor.Run(ctx, cmd)
-		if err == nil {
-			fmt.Fprintln(out, "Sidecar is healthy.")
-			return true
-		}
-		lastErr = err
-		fmt.Fprintf(out, "  attempt %d: %v\n", attempt, err)
-
-		if time.Now().After(deadline) {
-			fmt.Fprintf(out, "Warning: health check timed out (last error: %v). Services may still be starting.\n  Run: vibew deploy status --target <host> to check.\n  Run: vibew doctor --target <host> to diagnose.\n", lastErr)
-			return false
-		}
-
-		select {
-		case <-ctx.Done():
-			fmt.Fprintf(out, "Warning: health check cancelled (%v). Services may still be starting.\n  Run: vibew deploy status --target <host> to check.\n  Run: vibew doctor --target <host> to diagnose.\n", ctx.Err())
-			return false
-		case <-time.After(healthCheckInterval):
-		}
-	}
-}
-
-// healthCheckURL returns the health check URL using HTTPS when TLS is enabled
-// and HTTP otherwise. The host is always localhost because the probe runs on
-// the remote server itself via SSH.
-func healthCheckURL(port int, tlsEnabled bool) string {
-	scheme := "http"
-	if tlsEnabled {
-		scheme = "https"
-	}
-	return fmt.Sprintf("%s://localhost:%d/_vibewarden/health", scheme, port)
-}
-
-// healthCheckCmd returns the curl command used to probe the health endpoint.
-// When TLS is enabled it tries HTTPS first (with -k to skip cert verification),
-// falling back to HTTP if the TLS handshake fails (e.g., during ACME cert
-// acquisition). This prevents false health check failures during the first
-// minute of a deploy when the cert hasn't been issued yet.
-func healthCheckCmd(port int, tlsEnabled bool) string {
-	if tlsEnabled {
-		return fmt.Sprintf("curl -sfk https://localhost:%d/_vibewarden/health 2>/dev/null || curl -sf http://localhost:%d/_vibewarden/health", port, port)
-	}
-	return fmt.Sprintf("curl -sf http://localhost:%d/_vibewarden/health", port)
-}
-
 // ProjectNameFromConfig derives a project name from the config file path.
-// It returns the base name of the directory containing the config file, which
-// is the project directory name by convention. It is exported so the CLI can
-// compute the remote directory before calling Deploy.
+// It returns the base name of the directory containing the config file,
+// sanitised so spaces and dots become dashes.
 //
-// When configPath is empty the current working directory name is used as the
-// project name. When configPath is relative (e.g. "vibewarden.prod.yaml") it is
-// resolved to an absolute path first so that filepath.Dir returns the real
-// directory rather than ".".
+// When configPath is empty the current working directory name is used. When
+// configPath is relative it is resolved to an absolute path first so that
+// filepath.Dir returns the real directory rather than ".".
 func ProjectNameFromConfig(configPath string) string {
 	if configPath == "" {
 		wd, err := os.Getwd()
@@ -717,7 +77,6 @@ func ProjectNameFromConfig(configPath string) string {
 
 	dir := filepath.Dir(abs)
 	name := filepath.Base(dir)
-	// Sanitise: replace spaces and dots with dashes.
 	name = strings.ReplaceAll(name, " ", "-")
 	name = strings.ReplaceAll(name, ".", "-")
 	if name == "" || name == "." {
