--- internal/app/deploy/service_test.go
@@ -1,1650 +1,0 @@
-package deploy_test
-
-import (
-	"bytes"
-	"context"
-	"errors"
-	"fmt"
-	"io"
-	"strings"
-	"testing"
-
-	deployapp "github.com/vibewarden/vibewarden/internal/app/deploy"
-	"github.com/vibewarden/vibewarden/internal/config"
-	"github.com/vibewarden/vibewarden/internal/ports"
-)
-
-// fakeExecutor is a test double for ports.RemoteExecutor.
-type fakeExecutor struct {
-	// runResponses maps the command string to (output, error).
-	runResponses map[string]runResponse
-	// transferErr is returned by every Transfer call if set.
-	transferErr error
-	// transferExcludingErr is returned by every TransferExcluding call if set.
-	transferExcludingErr error
-	// transferFileErr is returned by every TransferFile call if set.
-	transferFileErr error
-	// dryRunChanges is returned by DryRunTransfer when non-nil.
-	dryRunChanges []string
-	// dryRunErr is returned by DryRunTransfer when non-nil.
-	dryRunErr error
-	// runCalls records every Run invocation for assertions.
-	runCalls []string
-	// transferCalls records every Transfer invocation for assertions.
-	transferCalls []transferCall
-	// transferExcludingCalls records every TransferExcluding invocation for assertions.
-	transferExcludingCalls []transferExcludingCall
-	// transferFileCalls records every TransferFile invocation for assertions.
-	transferFileCalls []transferFileCall
-}
-
-type runResponse struct {
-	output string
-	err    error
-}
-
-type transferCall struct {
-	localDir    string
-	remoteDir   string
-	deleteExtra bool
-}
-
-type transferExcludingCall struct {
-	localDir    string
-	remoteDir   string
-	deleteExtra bool
-	excludes    []string
-}
-
-type transferFileCall struct {
-	localFile  string
-	remotePath string
-}
-
-func (f *fakeExecutor) Run(_ context.Context, cmd string) (string, error) {
-	f.runCalls = append(f.runCalls, cmd)
-	if r, ok := f.runResponses[cmd]; ok {
-		return r.output, r.err
-	}
-	// Default: success with empty output.
-	return "", nil
-}
-
-func (f *fakeExecutor) RunStream(_ context.Context, cmd string, stdout, _ io.Writer) error {
-	f.runCalls = append(f.runCalls, cmd)
-	if r, ok := f.runResponses[cmd]; ok {
-		if r.output != "" {
-			fmt.Fprint(stdout, r.output)
-		}
-		return r.err
-	}
-	return nil
-}
-
-func (f *fakeExecutor) Transfer(_ context.Context, localDir, remoteDir string, deleteExtra bool) error {
-	f.transferCalls = append(f.transferCalls, transferCall{localDir: localDir, remoteDir: remoteDir, deleteExtra: deleteExtra})
-	return f.transferErr
-}
-
-func (f *fakeExecutor) TransferExcluding(_ context.Context, localDir, remoteDir string, deleteExtra bool, excludes []string) error {
-	f.transferExcludingCalls = append(f.transferExcludingCalls, transferExcludingCall{
-		localDir: localDir, remoteDir: remoteDir, deleteExtra: deleteExtra, excludes: excludes,
-	})
-	return f.transferExcludingErr
-}
-
-func (f *fakeExecutor) TransferFile(_ context.Context, localFile, remotePath string) error {
-	f.transferFileCalls = append(f.transferFileCalls, transferFileCall{localFile: localFile, remotePath: remotePath})
-	return f.transferFileErr
-}
-
-func (f *fakeExecutor) DryRunTransfer(_ context.Context, _, _ string) ([]string, error) {
-	if f.dryRunErr != nil {
-		return nil, f.dryRunErr
-	}
-	return f.dryRunChanges, nil
-}
-
-// fakeGenerator is a test double for ports.ConfigGenerator.
-type fakeGenerator struct {
-	err error
-}
-
-func (f *fakeGenerator) Generate(_ context.Context, _ ports.GeneratorInput, _ string) error {
-	return f.err
-}
-
-// fakeImageExporter is a test double for ports.ImageExporter.
-type fakeImageExporter struct {
-	err error
-}
-
-func (f *fakeImageExporter) Save(_ context.Context, _, _ string) error {
-	return f.err
-}
-
-// defaultConfig returns a minimal Config for testing.
-func defaultConfig() *config.Config {
-	return &config.Config{
-		Server: config.ServerConfig{Port: 8443},
-	}
-}
-
-func TestService_Deploy_HappyPath(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{}
-
-	// fakeExecutor returns success (nil error) for any unrecognised command,
-	// including the curl health-check probe.
-	svc := deployapp.NewService(executor, generator)
-
-	var buf bytes.Buffer
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/myproject/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-		Out:          &buf,
-	})
-	if err != nil {
-		t.Fatalf("Deploy() unexpected error: %v\noutput:\n%s", err, buf.String())
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "Deploy complete") {
-		t.Errorf("expected 'Deploy complete' in output, got:\n%s", out)
-	}
-
-	// Verify that prerequisite checks ran.
-	assertRunCalled(t, executor.runCalls, "which docker")
-	assertRunCalled(t, executor.runCalls, "docker compose version")
-
-	// Verify transfer was called (the deploy bundle rsync).
-	if len(executor.transferCalls) == 0 {
-		t.Error("expected Transfer to be called at least once")
-	}
-
-	// Verify docker compose up was called.
-	assertRunCalledContains(t, executor.runCalls, "docker compose up -d")
-}
-
-func TestService_Deploy_BundleFails(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{err: errors.New("template error")}
-
-	svc := deployapp.NewService(executor, generator)
-
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err == nil {
-		t.Fatal("expected error when generator fails")
-	}
-	if !strings.Contains(err.Error(), "generating config files") {
-		t.Errorf("error message should mention 'generating config files', got: %v", err)
-	}
-}
-
-func TestService_Deploy_MissingDocker(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"which docker": {err: errors.New("not found")},
-		},
-	}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err == nil {
-		t.Fatal("expected error when docker is missing")
-	}
-	if !strings.Contains(err.Error(), "remote prerequisites") {
-		t.Errorf("error should mention 'remote prerequisites', got: %v", err)
-	}
-}
-
-func TestService_Deploy_TransferFails(t *testing.T) {
-	executor := &fakeExecutor{
-		transferErr: errors.New("rsync failed"),
-	}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err == nil {
-		t.Fatal("expected error when transfer fails")
-	}
-	if !strings.Contains(err.Error(), "transferring") {
-		t.Errorf("error should mention 'transferring', got: %v", err)
-	}
-}
-
-// TestService_Deploy_HealthCheckTimeout verifies that a health-check timeout
-// DOES fail the deploy -- it returns ErrHealthCheck and prints an actionable
-// message. The operator can run "vibew deploy status" or "vibew doctor" to
-// diagnose.
-func TestService_Deploy_HealthCheckTimeout(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			// curl exits non-zero -> sidecar not yet healthy.
-			"curl -sf http://localhost:8443/_vibewarden/health": {err: errors.New("exit status 7")},
-		},
-	}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	// Use a cancelled context to avoid the full 60 s timeout.
-	ctx, cancel := context.WithCancel(context.Background())
-
-	// Cancel after the first health poll attempt completes.
-	go func() {
-		cancel()
-	}()
-
-	var buf bytes.Buffer
-	err := svc.Deploy(ctx, defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-		Out:          &buf,
-	})
-	// Health check timeout must now fail the deploy.
-	if !errors.Is(err, deployapp.ErrHealthCheck) {
-		t.Fatalf("Deploy() should return ErrHealthCheck when health check times out, got: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "health check failed") {
-		t.Errorf("expected 'health check failed' in output, got:\n%s", out)
-	}
-	if !strings.Contains(out, "Warning") {
-		t.Errorf("expected a warning message in output after health-check timeout, got:\n%s", out)
-	}
-	if !strings.Contains(out, "vibew doctor") {
-		t.Errorf("expected 'vibew doctor' suggestion in output, got:\n%s", out)
-	}
-}
-
-func TestService_Status(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/myproject/ ps": {output: "NAME   SERVICE   STATUS\nvw    vibewarden   running"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.Status(context.Background(), deployapp.StatusOptions{
-		ProjectName: "myproject",
-		Out:         &buf,
-	})
-	if err != nil {
-		t.Fatalf("Status() unexpected error: %v", err)
-	}
-
-	if !strings.Contains(buf.String(), "running") {
-		t.Errorf("expected status output to contain 'running', got:\n%s", buf.String())
-	}
-}
-
-func TestService_Status_DerivedFromConfigPath(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/myapp/ ps": {output: "NAME   SERVICE   STATUS\nvw    vibewarden   running"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.Status(context.Background(), deployapp.StatusOptions{
-		ConfigPath: "/home/user/myapp/vibewarden.yaml",
-		Out:        &buf,
-	})
-	if err != nil {
-		t.Fatalf("Status() unexpected error: %v", err)
-	}
-	if !strings.Contains(buf.String(), "running") {
-		t.Errorf("expected status output to contain 'running', got:\n%s", buf.String())
-	}
-}
-
-func TestService_Status_Error(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/vibewarden/ ps": {err: errors.New("connection refused")},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	// Use an explicit ProjectName so the test is not affected by the cwd used
-	// when resolving an empty ConfigPath.
-	err := svc.Status(context.Background(), deployapp.StatusOptions{ProjectName: "vibewarden"})
-	if err == nil {
-		t.Fatal("expected error when executor fails")
-	}
-	if !strings.Contains(err.Error(), "fetching remote status") {
-		t.Errorf("error should mention 'fetching remote status', got: %v", err)
-	}
-}
-
-func TestService_Logs(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/myproject/ logs --tail=50": {output: "log line 1\nlog line 2"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.Logs(context.Background(), deployapp.LogsOptions{
-		ProjectName: "myproject",
-		Lines:       50,
-		Out:         &buf,
-	})
-	if err != nil {
-		t.Fatalf("Logs() unexpected error: %v", err)
-	}
-
-	if !strings.Contains(buf.String(), "log line 1") {
-		t.Errorf("expected log output to contain 'log line 1', got:\n%s", buf.String())
-	}
-}
-
-func TestService_Logs_DerivedFromConfigPath(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/myapp/ logs --tail=20": {output: "log entry"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.Logs(context.Background(), deployapp.LogsOptions{
-		ConfigPath: "/home/user/myapp/vibewarden.yaml",
-		Lines:      20,
-		Out:        &buf,
-	})
-	if err != nil {
-		t.Fatalf("Logs() unexpected error: %v", err)
-	}
-	if !strings.Contains(buf.String(), "log entry") {
-		t.Errorf("expected 'log entry' in output, got:\n%s", buf.String())
-	}
-}
-
-func TestService_Logs_AllLines(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/myproject/ logs": {output: "all logs"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.Logs(context.Background(), deployapp.LogsOptions{
-		ProjectName: "myproject",
-		Lines:       0,
-		Out:         &buf,
-	})
-	if err != nil {
-		t.Fatalf("Logs() unexpected error: %v", err)
-	}
-	if !strings.Contains(buf.String(), "all logs") {
-		t.Errorf("expected 'all logs' in output, got:\n%s", buf.String())
-	}
-}
-
-func TestService_Logs_Error(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/myproject/ logs --tail=50": {err: errors.New("remote error")},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	err := svc.Logs(context.Background(), deployapp.LogsOptions{
-		ProjectName: "myproject",
-		Lines:       50,
-	})
-	if err == nil {
-		t.Fatal("expected error when executor fails")
-	}
-	if !strings.Contains(err.Error(), "fetching remote logs") {
-		t.Errorf("error should mention 'fetching remote logs', got: %v", err)
-	}
-}
-
-func TestService_Logs_Follow(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/myproject/ logs --tail=20 -f": {output: "streamed line 1\nstreamed line 2"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.Logs(context.Background(), deployapp.LogsOptions{
-		ProjectName: "myproject",
-		Lines:       20,
-		Follow:      true,
-		Out:         &buf,
-	})
-	if err != nil {
-		t.Fatalf("Logs(Follow=true) unexpected error: %v", err)
-	}
-
-	if !strings.Contains(buf.String(), "streamed line 1") {
-		t.Errorf("expected streamed output in buf, got:\n%s", buf.String())
-	}
-	// Verify RunStream was called (not Run) by checking the recorded command.
-	if len(executor.runCalls) != 1 {
-		t.Fatalf("expected exactly 1 run call, got: %v", executor.runCalls)
-	}
-	if !strings.Contains(executor.runCalls[0], "-f") {
-		t.Errorf("expected '-f' in the run call command, got: %q", executor.runCalls[0])
-	}
-}
-
-func TestService_Logs_Follow_NoLines(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/myproject/ logs -f": {output: "all streamed"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.Logs(context.Background(), deployapp.LogsOptions{
-		ProjectName: "myproject",
-		Lines:       0,
-		Follow:      true,
-		Out:         &buf,
-	})
-	if err != nil {
-		t.Fatalf("Logs(Follow=true, Lines=0) unexpected error: %v", err)
-	}
-	if !strings.Contains(buf.String(), "all streamed") {
-		t.Errorf("expected 'all streamed' in output, got:\n%s", buf.String())
-	}
-}
-
-func TestService_Logs_Follow_Error(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/myproject/ logs -f": {err: errors.New("stream broken")},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	err := svc.Logs(context.Background(), deployapp.LogsOptions{
-		ProjectName: "myproject",
-		Follow:      true,
-	})
-	if err == nil {
-		t.Fatal("expected error when RunStream fails")
-	}
-	if !strings.Contains(err.Error(), "streaming remote logs") {
-		t.Errorf("error should mention 'streaming remote logs', got: %v", err)
-	}
-}
-
-func TestService_Deploy_RemoteDir(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/home/user/myproject/vibewarden.prod.yaml",
-		ProjectName:  "myproject",
-		GeneratedDir: t.TempDir(),
-	})
-	if err != nil {
-		t.Fatalf("Deploy() unexpected error: %v", err)
-	}
-
-	// Verify the remote directory includes the project name.
-	found := false
-	for _, call := range executor.runCalls {
-		if strings.Contains(call, "~/vibewarden/myproject/") {
-			found = true
-			break
-		}
-	}
-	if !found {
-		t.Errorf("expected remote dir ~/vibewarden/myproject/ to appear in run calls, got: %v", executor.runCalls)
-	}
-}
-
-func TestService_Deploy_DockerComposeMissing(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"which docker":           {output: "/usr/bin/docker"},
-			"docker compose version": {err: errors.New("docker compose not found")},
-		},
-	}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err == nil {
-		t.Fatal("expected error when docker compose is missing")
-	}
-	if !strings.Contains(err.Error(), "remote prerequisites") {
-		t.Errorf("error should mention 'remote prerequisites', got: %v", err)
-	}
-}
-
-// assertRunCalled checks that cmd appears in calls.
-func assertRunCalled(t *testing.T, calls []string, cmd string) {
-	t.Helper()
-	for _, c := range calls {
-		if c == cmd {
-			return
-		}
-	}
-	t.Errorf("expected Run(%q) to be called, actual calls: %v", cmd, calls)
-}
-
-// assertRunCalledContains checks that at least one call contains the substring.
-func assertRunCalledContains(t *testing.T, calls []string, substr string) {
-	t.Helper()
-	for _, c := range calls {
-		if strings.Contains(c, substr) {
-			return
-		}
-	}
-	t.Errorf("expected a Run call containing %q, actual calls: %v", substr, calls)
-}
-
-func TestProjectNameFromConfig(t *testing.T) {
-	// We test the exported behaviour indirectly through Deploy's remote dir.
-	tests := []struct {
-		name        string
-		configPath  string
-		projectName string // explicit override
-		wantDirSub  string // expected substring in remote dir run calls
-	}{
-		{
-			name:       "derive from config dir",
-			configPath: "/home/user/myapp/vibewarden.yaml",
-			wantDirSub: "~/vibewarden/myapp/",
-		},
-		{
-			name:        "explicit project name overrides config dir",
-			configPath:  "/home/user/myapp/vibewarden.yaml",
-			projectName: "production",
-			wantDirSub:  "~/vibewarden/production/",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			executor := &fakeExecutor{}
-			generator := &fakeGenerator{}
-
-			svc := deployapp.NewService(executor, generator)
-
-			_ = svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-				ConfigPath:   tt.configPath,
-				ProjectName:  tt.projectName,
-				GeneratedDir: t.TempDir(),
-			})
-
-			found := false
-			for _, c := range executor.runCalls {
-				if strings.Contains(c, tt.wantDirSub) {
-					found = true
-					break
-				}
-			}
-			if !found {
-				t.Errorf("expected remote dir %q in run calls, got: %v", tt.wantDirSub, executor.runCalls)
-			}
-		})
-	}
-}
-
-// TestService_Deploy_NilOut ensures a nil Out does not panic.
-func TestService_Deploy_NilOut(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{err: errors.New("stop early")}
-
-	svc := deployapp.NewService(executor, generator)
-
-	// Should not panic even with nil Out.
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-		Out:          nil,
-	})
-	if err == nil {
-		t.Fatal("expected error")
-	}
-}
-
-// TestService_Deploy_ComposeUpFails verifies that a compose up failure is surfaced.
-func TestService_Deploy_ComposeUpFails(t *testing.T) {
-	executor := &mockRunExecutor{
-		runFn: func(cmd string) (string, error) {
-			if strings.Contains(cmd, "docker compose up -d") {
-				return "", errors.New("compose up failed")
-			}
-			return "", nil
-		},
-	}
-
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err == nil {
-		t.Fatal("expected error when compose up fails")
-	}
-	if !strings.Contains(err.Error(), "docker compose up") {
-		t.Errorf("error should mention 'docker compose up', got: %v", err)
-	}
-}
-
-// mockRunExecutor is a flexible test double with a custom Run function.
-type mockRunExecutor struct {
-	runFn                  func(cmd string) (string, error)
-	dryRunFn               func(localDir, remoteDir string) ([]string, error)
-	transferCalls          []transferCall
-	transferExcludingCalls []transferExcludingCall
-	transferFileCalls      []transferFileCall
-}
-
-func (m *mockRunExecutor) Run(_ context.Context, cmd string) (string, error) {
-	if m.runFn != nil {
-		return m.runFn(cmd)
-	}
-	return "", nil
-}
-
-func (m *mockRunExecutor) RunStream(_ context.Context, _ string, _ io.Writer, _ io.Writer) error {
-	return nil
-}
-
-func (m *mockRunExecutor) Transfer(_ context.Context, localDir, remoteDir string, deleteExtra bool) error {
-	m.transferCalls = append(m.transferCalls, transferCall{localDir: localDir, remoteDir: remoteDir, deleteExtra: deleteExtra})
-	return nil
-}
-
-func (m *mockRunExecutor) TransferExcluding(_ context.Context, localDir, remoteDir string, deleteExtra bool, excludes []string) error {
-	m.transferExcludingCalls = append(m.transferExcludingCalls, transferExcludingCall{
-		localDir: localDir, remoteDir: remoteDir, deleteExtra: deleteExtra, excludes: excludes,
-	})
-	return nil
-}
-
-func (m *mockRunExecutor) TransferFile(_ context.Context, localFile, remotePath string) error {
-	m.transferFileCalls = append(m.transferFileCalls, transferFileCall{localFile: localFile, remotePath: remotePath})
-	return nil
-}
-
-func (m *mockRunExecutor) DryRunTransfer(_ context.Context, localDir, remoteDir string) ([]string, error) {
-	if m.dryRunFn != nil {
-		return m.dryRunFn(localDir, remoteDir)
-	}
-	return nil, nil
-}
-
-// TestService_Deploy_ImageMode verifies that when cfg.App.Build is empty the
-// deploy flow runs docker compose pull followed by docker compose up -d.
-func TestService_Deploy_ImageMode(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	cfg := defaultConfig()
-	cfg.App.Image = "myapp:latest"
-
-	err := svc.Deploy(context.Background(), cfg, deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err != nil {
-		t.Fatalf("Deploy() unexpected error: %v", err)
-	}
-
-	// docker compose pull must be called.
-	assertRunCalledContains(t, executor.runCalls, "docker compose pull")
-	// docker compose up -d (without --build) must be called.
-	assertRunCalledContains(t, executor.runCalls, "docker compose up -d")
-
-	// docker compose up -d --build must NOT be called.
-	for _, c := range executor.runCalls {
-		if strings.Contains(c, "--build") {
-			t.Errorf("did not expect --build flag in image mode, but got run call: %q", c)
-		}
-	}
-
-	// Only the bundle transfer is expected (one Transfer call).
-	if len(executor.transferCalls) != 1 {
-		t.Errorf("expected exactly 1 Transfer call in image mode, got %d: %v", len(executor.transferCalls), executor.transferCalls)
-	}
-}
-
-// TestService_Deploy_BuildMode verifies that when cfg.App.Build is set the
-// deploy flow transfers the image (not source code) and runs docker compose
-// up -d without --build.
-func TestService_Deploy_BuildMode(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	cfg := defaultConfig()
-	cfg.App.Build = "."
-
-	err := svc.Deploy(context.Background(), cfg, deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err != nil {
-		t.Fatalf("Deploy() unexpected error: %v", err)
-	}
-
-	// docker compose up -d must be called (without --build).
-	assertRunCalledContains(t, executor.runCalls, "docker compose up -d")
-
-	// --build must NOT appear -- image is pre-built and transferred.
-	for _, c := range executor.runCalls {
-		if strings.Contains(c, "--build") {
-			t.Errorf("did not expect --build flag in image transfer mode, got: %q", c)
-		}
-	}
-
-	// docker compose pull vibewarden (sidecar image refresh) must be called.
-	assertRunCalledContains(t, executor.runCalls, "docker compose pull vibewarden")
-
-	// Only the bundle transfer is expected (one Transfer call).
-	if len(executor.transferCalls) != 1 {
-		t.Errorf("expected 1 Transfer call in build mode (bundle only), got %d: %v",
-			len(executor.transferCalls), executor.transferCalls)
-	}
-
-	// No TransferExcluding -- no source code transfer.
-	if len(executor.transferExcludingCalls) != 0 {
-		t.Errorf("expected 0 TransferExcluding calls in build mode, got %d: %v",
-			len(executor.transferExcludingCalls), executor.transferExcludingCalls)
-	}
-}
-
-// TestService_Deploy_BuildMode_ImageTransferFails verifies that a failure to
-// transfer the locally-built image is propagated correctly.
-func TestService_Deploy_BuildMode_ImageTransferFails(t *testing.T) {
-	executor := &fakeExecutor{
-		transferFileErr: fmt.Errorf("rsync failed"),
-	}
-	generator := &fakeGenerator{}
-
-	exporter := &fakeImageExporter{}
-	svc := deployapp.NewService(executor, generator).WithImageExporter(exporter)
-
-	cfg := defaultConfig()
-	cfg.App.Build = "."
-
-	err := svc.Deploy(context.Background(), cfg, deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err == nil {
-		t.Fatal("expected error when image transfer fails")
-	}
-	if !strings.Contains(err.Error(), "transferring") {
-		t.Errorf("error should mention 'transferring', got: %v", err)
-	}
-}
-
-// TestProjectNameFromConfig_Unit directly exercises ProjectNameFromConfig with
-// various inputs, including relative paths that previously returned "." and
-// fell back to "vibewarden" instead of the actual directory name.
-func TestProjectNameFromConfig_Unit(t *testing.T) {
-	tests := []struct {
-		name       string
-		configPath string
-		want       string
-	}{
-		{
-			name:       "absolute path uses directory basename",
-			configPath: "/home/user/myapp/vibewarden.yaml",
-			want:       "myapp",
-		},
-		{
-			name:       "absolute path with dots in dir sanitised to dashes",
-			configPath: "/home/user/my.app/vibewarden.yaml",
-			want:       "my-app",
-		},
-		{
-			name:       "absolute path with spaces in dir sanitised to dashes",
-			configPath: "/home/user/my app/vibewarden.yaml",
-			want:       "my-app",
-		},
-		{
-			name:       "relative filename only resolves to cwd basename",
-			configPath: "vibewarden.prod.yaml",
-			// Cannot assert the exact name (depends on cwd) but must not be "vibewarden".
-			// Tested via the want field set dynamically below.
-			want: "", // sentinel: see special handling below
-		},
-		{
-			name:       "empty path resolves to cwd basename",
-			configPath: "",
-			want:       "", // sentinel
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := deployapp.ProjectNameFromConfig(tt.configPath)
-
-			if tt.want == "" {
-				// Relative/empty path -- just assert we don't get "." or empty string.
-				if got == "." || got == "" {
-					t.Errorf("ProjectNameFromConfig(%q) = %q, want a non-empty non-dot name", tt.configPath, got)
-				}
-				return
-			}
-
-			if got != tt.want {
-				t.Errorf("ProjectNameFromConfig(%q) = %q, want %q", tt.configPath, got, tt.want)
-			}
-		})
-	}
-}
-
-// TestProjectNameFromConfig_RelativeDoesNotReturnDot ensures that a plain
-// filename (no directory component) does not produce "." after cleaning.
-// This was the root bug: filepath.Dir("vibewarden.prod.yaml") == "." and
-// filepath.Base(".") == ".", causing a fallback to "vibewarden" instead of the
-// actual working directory name.
-func TestProjectNameFromConfig_RelativeDoesNotReturnDot(t *testing.T) {
-	name := deployapp.ProjectNameFromConfig("vibewarden.prod.yaml")
-	if name == "." {
-		t.Errorf("ProjectNameFromConfig returned \".\", expected the cwd basename")
-	}
-	if name == "" {
-		t.Errorf("ProjectNameFromConfig returned empty string")
-	}
-}
-
-func TestService_ListSites(t *testing.T) {
-	tests := []struct {
-		name     string
-		executor *fakeExecutor
-		want     []string
-		wantErr  bool
-	}{
-		{
-			name: "lists multiple sites",
-			executor: &fakeExecutor{
-				runResponses: map[string]runResponse{
-					"ls -1 ~/vibewarden/sites/ 2>/dev/null || true": {output: "blog\napi\nshop"},
-				},
-			},
-			want: []string{"blog", "api", "shop"},
-		},
-		{
-			name: "empty directory returns nil",
-			executor: &fakeExecutor{
-				runResponses: map[string]runResponse{
-					"ls -1 ~/vibewarden/sites/ 2>/dev/null || true": {output: ""},
-				},
-			},
-			want: nil,
-		},
-		{
-			name: "missing directory returns nil via || true",
-			executor: &fakeExecutor{
-				runResponses: map[string]runResponse{
-					"ls -1 ~/vibewarden/sites/ 2>/dev/null || true": {output: ""},
-				},
-			},
-			want: nil,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			svc := deployapp.NewService(tt.executor, nil)
-			got, err := svc.ListSites(context.Background())
-			if (err != nil) != tt.wantErr {
-				t.Fatalf("ListSites() error = %v, wantErr %v", err, tt.wantErr)
-			}
-			if len(got) != len(tt.want) {
-				t.Fatalf("ListSites() returned %d sites, want %d", len(got), len(tt.want))
-			}
-			for i, name := range got {
-				if name != tt.want[i] {
-					t.Errorf("ListSites()[%d] = %q, want %q", i, name, tt.want[i])
-				}
-			}
-		})
-	}
-}
-
-func TestService_StatusMultiApp_SpecificApp(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/sites/blog/ ps": {output: "blog container running"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.StatusMultiApp(context.Background(), "blog", &buf)
-	if err != nil {
-		t.Fatalf("StatusMultiApp() unexpected error: %v", err)
-	}
-	if !strings.Contains(buf.String(), "=== Site: blog ===") {
-		t.Errorf("expected site header in output, got:\n%s", buf.String())
-	}
-	if !strings.Contains(buf.String(), "blog container running") {
-		t.Errorf("expected site status in output, got:\n%s", buf.String())
-	}
-}
-
-func TestService_StatusMultiApp_AllSites(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/.sidecar/ ps":   {output: "sidecar running"},
-			"ls -1 ~/vibewarden/sites/ 2>/dev/null || true":                  {output: "app1\napp2"},
-			"docker compose --project-directory ~/vibewarden/sites/app1/ ps": {output: "app1 running"},
-			"docker compose --project-directory ~/vibewarden/sites/app2/ ps": {output: "app2 running"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.StatusMultiApp(context.Background(), "", &buf)
-	if err != nil {
-		t.Fatalf("StatusMultiApp() unexpected error: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "=== Sidecar ===") {
-		t.Errorf("expected sidecar header in output, got:\n%s", out)
-	}
-	if !strings.Contains(out, "=== Site: app1 ===") {
-		t.Errorf("expected app1 header in output, got:\n%s", out)
-	}
-	if !strings.Contains(out, "=== Site: app2 ===") {
-		t.Errorf("expected app2 header in output, got:\n%s", out)
-	}
-}
-
-func TestService_StatusMultiApp_Error(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/sites/bad/ ps": {err: errors.New("not found")},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	err := svc.StatusMultiApp(context.Background(), "bad", io.Discard)
-	if err == nil {
-		t.Fatal("expected error when fetching status for a bad site")
-	}
-	if !strings.Contains(err.Error(), "fetching status for site") {
-		t.Errorf("error should mention site, got: %v", err)
-	}
-}
-
-func TestService_LogsMultiApp_SpecificApp(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/sites/blog/ logs --tail=20": {output: "blog log line"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.LogsMultiApp(context.Background(), "blog", deployapp.LogsOptions{
-		Lines: 20,
-		Out:   &buf,
-	})
-	if err != nil {
-		t.Fatalf("LogsMultiApp() unexpected error: %v", err)
-	}
-	if !strings.Contains(buf.String(), "blog log line") {
-		t.Errorf("expected blog logs in output, got:\n%s", buf.String())
-	}
-}
-
-func TestService_LogsMultiApp_Sidecar(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/.sidecar/ logs --tail=50": {output: "sidecar log"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.LogsMultiApp(context.Background(), "", deployapp.LogsOptions{
-		Lines: 50,
-		Out:   &buf,
-	})
-	if err != nil {
-		t.Fatalf("LogsMultiApp() unexpected error: %v", err)
-	}
-	if !strings.Contains(buf.String(), "sidecar log") {
-		t.Errorf("expected sidecar logs in output, got:\n%s", buf.String())
-	}
-}
-
-func TestService_LogsMultiApp_Follow(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/sites/api/ logs --tail=10 -f": {output: "streaming"},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	var buf bytes.Buffer
-	err := svc.LogsMultiApp(context.Background(), "api", deployapp.LogsOptions{
-		Lines:  10,
-		Follow: true,
-		Out:    &buf,
-	})
-	if err != nil {
-		t.Fatalf("LogsMultiApp() unexpected error: %v", err)
-	}
-	if !strings.Contains(buf.String(), "streaming") {
-		t.Errorf("expected streaming output, got:\n%s", buf.String())
-	}
-}
-
-func TestService_LogsMultiApp_Error(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			"docker compose --project-directory ~/vibewarden/sites/bad/ logs --tail=50": {err: errors.New("not found")},
-		},
-	}
-
-	svc := deployapp.NewService(executor, nil)
-
-	err := svc.LogsMultiApp(context.Background(), "bad", deployapp.LogsOptions{
-		Lines: 50,
-	})
-	if err == nil {
-		t.Fatal("expected error when fetching logs for a bad site")
-	}
-	if !strings.Contains(err.Error(), "fetching remote logs") {
-		t.Errorf("error should mention 'fetching remote logs', got: %v", err)
-	}
-}
-
-// Ensure fmt is used (used in assertRunCalledContains via Errorf).
-var _ = fmt.Sprintf
-
-// TestService_Deploy_PullsSidecarBeforeUp verifies that "docker compose pull
-// vibewarden" is always executed before starting services, so that
-// the sidecar image is never served from a stale Docker layer cache.
-func TestService_Deploy_PullsSidecarBeforeUp(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	cfg := defaultConfig()
-	cfg.App.Build = "."
-
-	err := svc.Deploy(context.Background(), cfg, deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err != nil {
-		t.Fatalf("Deploy() unexpected error: %v", err)
-	}
-
-	// "docker compose pull vibewarden" must appear in run calls.
-	assertRunCalledContains(t, executor.runCalls, "docker compose pull vibewarden")
-
-	// Verify ordering: pull vibewarden must come before compose up.
-	pullIdx, upIdx := -1, -1
-	for i, c := range executor.runCalls {
-		if strings.Contains(c, "docker compose pull vibewarden") && pullIdx == -1 {
-			pullIdx = i
-		}
-		if strings.Contains(c, "docker compose up -d") && upIdx == -1 {
-			upIdx = i
-		}
-	}
-	if pullIdx == -1 {
-		t.Error("'docker compose pull vibewarden' was not found in run calls")
-	}
-	if upIdx == -1 {
-		t.Error("'docker compose up -d' was not found in run calls")
-	}
-	if pullIdx != -1 && upIdx != -1 && pullIdx >= upIdx {
-		t.Errorf("expected 'docker compose pull vibewarden' (call %d) before 'docker compose up -d' (call %d)", pullIdx, upIdx)
-	}
-}
-
-// TestService_Deploy_PullsSidecarInImageMode verifies that "docker compose pull
-// vibewarden" is executed in image mode as well, before the full pull.
-func TestService_Deploy_PullsSidecarInImageMode(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	cfg := defaultConfig()
-	cfg.App.Image = "myapp:latest"
-
-	err := svc.Deploy(context.Background(), cfg, deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-	})
-	if err != nil {
-		t.Fatalf("Deploy() unexpected error: %v", err)
-	}
-
-	// Both the targeted sidecar pull and the full pull must be present.
-	assertRunCalledContains(t, executor.runCalls, "docker compose pull vibewarden")
-	assertRunCalledContains(t, executor.runCalls, "docker compose pull")
-}
-
-// TestService_Deploy_DriftDetected_AbortsWithoutForce verifies that Deploy
-// returns a DriftError when remote files have diverged and --force is not set.
-func TestService_Deploy_DriftDetected_AbortsWithoutForce(t *testing.T) {
-	executor := &fakeExecutor{
-		dryRunChanges: []string{
-			">f..T...... docker-compose.yml",
-			"*deleting   custom-config.yml",
-		},
-	}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-		Force:        false,
-	})
-	if err == nil {
-		t.Fatal("expected DriftError when drift is detected without --force")
-	}
-
-	var driftErr *deployapp.DriftError
-	if !errors.As(err, &driftErr) {
-		t.Fatalf("expected error to be *DriftError, got %T: %v", err, err)
-	}
-	if len(driftErr.Changes) != 2 {
-		t.Errorf("expected 2 changes, got %d", len(driftErr.Changes))
-	}
-	if !strings.Contains(err.Error(), "docker-compose.yml") {
-		t.Errorf("error message should list affected files, got: %v", err)
-	}
-	if !strings.Contains(err.Error(), "vibew deploy --force") {
-		t.Errorf("error message should suggest --force, got: %v", err)
-	}
-}
-
-// TestService_Deploy_DriftDetected_ProceedsWithForce verifies that Deploy
-// ignores drift when --force is set and completes successfully.
-func TestService_Deploy_DriftDetected_ProceedsWithForce(t *testing.T) {
-	executor := &fakeExecutor{
-		dryRunChanges: []string{
-			">f..T...... docker-compose.yml",
-		},
-	}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	var buf bytes.Buffer
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-		Force:        true,
-		Out:          &buf,
-	})
-	if err != nil {
-		t.Fatalf("Deploy() with --force should succeed even with drift, got: %v", err)
-	}
-	if !strings.Contains(buf.String(), "Deploy complete") {
-		t.Errorf("expected 'Deploy complete' in output, got:\n%s", buf.String())
-	}
-}
-
-// TestService_Deploy_NoDrift_ProceedsWithoutForce verifies that Deploy proceeds
-// normally when no drift is detected and --force is not set.
-func TestService_Deploy_NoDrift_ProceedsWithoutForce(t *testing.T) {
-	executor := &fakeExecutor{
-		dryRunChanges: nil, // no changes
-	}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	var buf bytes.Buffer
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-		Force:        false,
-		Out:          &buf,
-	})
-	if err != nil {
-		t.Fatalf("Deploy() without drift should succeed, got: %v", err)
-	}
-	if !strings.Contains(buf.String(), "Deploy complete") {
-		t.Errorf("expected 'Deploy complete' in output, got:\n%s", buf.String())
-	}
-}
-
-// TestService_Deploy_DryRunFails_FallsThrough verifies that a dry-run failure
-// (e.g. first deploy when the remote dir does not exist) prints a note but
-// does not abort the deploy.
-func TestService_Deploy_DryRunFails_FallsThrough(t *testing.T) {
-	executor := &fakeExecutor{
-		dryRunErr: errors.New("rsync: connection refused"),
-	}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	var buf bytes.Buffer
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-		Force:        false,
-		Out:          &buf,
-	})
-	if err != nil {
-		t.Fatalf("Deploy() should succeed when dry-run fails (first deploy), got: %v", err)
-	}
-	if !strings.Contains(buf.String(), "drift detection skipped") {
-		t.Errorf("expected note about skipped drift detection, got:\n%s", buf.String())
-	}
-	if !strings.Contains(buf.String(), "Deploy complete") {
-		t.Errorf("expected 'Deploy complete' in output, got:\n%s", buf.String())
-	}
-}
-
-// TestDriftError_ErrorMessage verifies the DriftError message format.
-func TestDriftError_ErrorMessage(t *testing.T) {
-	err := &deployapp.DriftError{
-		Changes: []string{
-			">f..T...... docker-compose.yml",
-			"*deleting   custom-config.yml",
-		},
-	}
-
-	msg := err.Error()
-	if !strings.Contains(msg, "remote files have been modified") {
-		t.Errorf("expected 'remote files have been modified' in message, got: %s", msg)
-	}
-	if !strings.Contains(msg, "docker-compose.yml") {
-		t.Errorf("expected 'docker-compose.yml' in message, got: %s", msg)
-	}
-	if !strings.Contains(msg, "custom-config.yml") {
-		t.Errorf("expected 'custom-config.yml' in message, got: %s", msg)
-	}
-	if !strings.Contains(msg, "vibew deploy --force") {
-		t.Errorf("expected '--force' suggestion in message, got: %s", msg)
-	}
-}
-
-// TestDriftError_CategorisesDeployOwnedAndUserFiles verifies that DriftError
-// separates deploy-owned files from user-modified files in the error message.
-func TestDriftError_CategorisesDeployOwnedAndUserFiles(t *testing.T) {
-	tests := []struct {
-		name           string
-		changes        []string
-		wantDeployOwn  []string // substrings that should appear under "Deploy-owned"
-		wantUserMod    []string // substrings that should appear under "User-modified"
-		wantNoDeployHd bool     // true if "Deploy-owned" header should be absent
-		wantNoUserHd   bool     // true if "User-modified" header should be absent
-	}{
-		{
-			name: "mixed deploy-owned and user files",
-			changes: []string{
-				">f..T...... docker-compose.yml",
-				">f.s....... vibewarden.yaml",
-				"*deleting   custom-config.yml",
-				">f..T...... kratos/kratos.yml",
-			},
-			wantDeployOwn: []string{"docker-compose.yml", "vibewarden.yaml", "kratos/kratos.yml"},
-			wantUserMod:   []string{"custom-config.yml"},
-		},
-		{
-			name: "only deploy-owned files",
-			changes: []string{
-				">f..T...... .credentials",
-				">f..T...... .env",
-			},
-			wantDeployOwn: []string{".credentials", ".env"},
-			wantNoUserHd:  true,
-		},
-		{
-			name: "only user files",
-			changes: []string{
-				"*deleting   my-custom-script.sh",
-			},
-			wantUserMod:    []string{"my-custom-script.sh"},
-			wantNoDeployHd: true,
-		},
-		{
-			name: "openbao prefix is deploy-owned",
-			changes: []string{
-				">f..T...... openbao/config.hcl",
-			},
-			wantDeployOwn: []string{"openbao/config.hcl"},
-			wantNoUserHd:  true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := &deployapp.DriftError{Changes: tt.changes}
-			msg := err.Error()
-
-			for _, s := range tt.wantDeployOwn {
-				if !strings.Contains(msg, s) {
-					t.Errorf("expected %q in deploy-owned section, got:\n%s", s, msg)
-				}
-			}
-			for _, s := range tt.wantUserMod {
-				if !strings.Contains(msg, s) {
-					t.Errorf("expected %q in user-modified section, got:\n%s", s, msg)
-				}
-			}
-
-			if !tt.wantNoDeployHd {
-				if len(tt.wantDeployOwn) > 0 && !strings.Contains(msg, "Deploy-owned") {
-					t.Errorf("expected 'Deploy-owned' header in message, got:\n%s", msg)
-				}
-			} else if strings.Contains(msg, "Deploy-owned") {
-				t.Errorf("did not expect 'Deploy-owned' header in message, got:\n%s", msg)
-			}
-
-			if !tt.wantNoUserHd {
-				if len(tt.wantUserMod) > 0 && !strings.Contains(msg, "User-modified") {
-					t.Errorf("expected 'User-modified' header in message, got:\n%s", msg)
-				}
-			} else if strings.Contains(msg, "User-modified") {
-				t.Errorf("did not expect 'User-modified' header in message, got:\n%s", msg)
-			}
-
-			if !strings.Contains(msg, "vibew deploy --force") {
-				t.Errorf("expected '--force' suggestion in message, got:\n%s", msg)
-			}
-		})
-	}
-}
-
-// TestService_Deploy_CredentialPreservationBeforeDriftDetection verifies that
-// credential files are fetched from the remote BEFORE drift detection runs.
-// This prevents false-positive drift caused by freshly generated credentials
-// in the bundle differing from existing remote credentials. (#1031)
-func TestService_Deploy_CredentialPreservationBeforeDriftDetection(t *testing.T) {
-	// Track the order of operations to verify credential fetch precedes dry-run.
-	var opOrder []string
-
-	executor := &mockRunExecutor{
-		runFn: func(cmd string) (string, error) {
-			if strings.Contains(cmd, "test -f") && strings.Contains(cmd, ".credentials") {
-				opOrder = append(opOrder, "cred-test-.credentials")
-				return "", nil // file exists
-			}
-			if strings.Contains(cmd, "cat") && strings.Contains(cmd, ".credentials") {
-				opOrder = append(opOrder, "cred-cat-.credentials")
-				return "POSTGRES_PASSWORD=existing_password", nil
-			}
-			if strings.Contains(cmd, "test -f") && strings.Contains(cmd, ".env") {
-				opOrder = append(opOrder, "cred-test-.env")
-				return "", nil // file exists
-			}
-			if strings.Contains(cmd, "cat") && strings.Contains(cmd, ".env") {
-				opOrder = append(opOrder, "cred-cat-.env")
-				return "APP_SECRET=existing_secret", nil
-			}
-			return "", nil
-		},
-	}
-
-	// Override DryRunTransfer to record when it is called.
-	dryRunCalled := false
-	origDryRun := executor.dryRunFn
-	_ = origDryRun
-	executor.dryRunFn = func(_, _ string) ([]string, error) {
-		dryRunCalled = true
-		opOrder = append(opOrder, "dry-run")
-		return nil, nil // no drift
-	}
-
-	generator := &fakeGenerator{}
-	svc := deployapp.NewService(executor, generator)
-
-	var buf bytes.Buffer
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-		Force:        false,
-		Out:          &buf,
-	})
-	if err != nil {
-		t.Fatalf("Deploy() unexpected error: %v", err)
-	}
-
-	if !dryRunCalled {
-		t.Fatal("expected DryRunTransfer to be called (Force=false)")
-	}
-
-	// Verify ordering: credential operations must come before dry-run.
-	dryRunIdx := -1
-	lastCredIdx := -1
-	for i, op := range opOrder {
-		if op == "dry-run" && dryRunIdx == -1 {
-			dryRunIdx = i
-		}
-		if strings.HasPrefix(op, "cred-") {
-			lastCredIdx = i
-		}
-	}
-	if lastCredIdx == -1 {
-		t.Fatal("expected credential fetch operations to be recorded")
-	}
-	if dryRunIdx == -1 {
-		t.Fatal("expected dry-run operation to be recorded")
-	}
-	if lastCredIdx >= dryRunIdx {
-		t.Errorf("credential fetch (index %d) must happen before dry-run (index %d); operation order: %v",
-			lastCredIdx, dryRunIdx, opOrder)
-	}
-}
-
-// TestService_Deploy_HealthCheckWarnOnTimeout verifies the exact warning message
-// format when the health check times out, including the doctor suggestion.
-func TestService_Deploy_HealthCheckWarnOnTimeout(t *testing.T) {
-	executor := &fakeExecutor{
-		runResponses: map[string]runResponse{
-			// curl exits non-zero -> sidecar not yet healthy.
-			"curl -sf http://localhost:8443/_vibewarden/health": {err: errors.New("exit status 7")},
-		},
-	}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() { cancel() }()
-
-	var buf bytes.Buffer
-	err := svc.Deploy(ctx, defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		GeneratedDir: t.TempDir(),
-		Out:          &buf,
-	})
-
-	if !errors.Is(err, deployapp.ErrHealthCheck) {
-		t.Fatalf("Deploy() should return ErrHealthCheck on health-check timeout, got: %v", err)
-	}
-
-	out := buf.String()
-	if !strings.Contains(out, "Warning") {
-		t.Errorf("expected warning in output, got:\n%s", out)
-	}
-	if !strings.Contains(out, "vibew deploy status") {
-		t.Errorf("expected 'vibew deploy status' suggestion in output, got:\n%s", out)
-	}
-	if !strings.Contains(out, "vibew doctor") {
-		t.Errorf("expected 'vibew doctor' suggestion in output, got:\n%s", out)
-	}
-	// "Deploy complete" should NOT appear when health check fails.
-	if strings.Contains(out, "Deploy complete.") {
-		t.Errorf("'Deploy complete.' should NOT appear when health check fails, got:\n%s", out)
-	}
-}
-
-// TestService_Deploy_HealthCheckAlwaysUsesLocalhost verifies that the health
-// check always probes localhost via SSH, regardless of TLS configuration or
-// domain. When TLS is enabled, it uses https with -k; when disabled, plain http.
-// This avoids DNS propagation, firewall, and TLS certificate issuance
-// dependencies.
-func TestService_Deploy_HealthCheckAlwaysUsesLocalhost(t *testing.T) {
-	tests := []struct {
-		name        string
-		cfg         *config.Config
-		wantCurlCmd string
-	}{
-		{
-			name: "TLS disabled uses HTTP localhost",
-			cfg: &config.Config{
-				Server: config.ServerConfig{Port: 8080},
-			},
-			wantCurlCmd: "curl -sf http://localhost:8080/_vibewarden/health",
-		},
-		{
-			name: "TLS enabled without domain uses HTTPS with -k",
-			cfg: &config.Config{
-				Server: config.ServerConfig{Port: 8443},
-				TLS:    config.TLSConfig{Enabled: true},
-			},
-			wantCurlCmd: "curl -sfk https://localhost:8443/_vibewarden/health 2>/dev/null || curl -sf http://localhost:8443/_vibewarden/health",
-		},
-		{
-			name: "TLS enabled with domain uses HTTPS with -k on localhost",
-			cfg: &config.Config{
-				Server: config.ServerConfig{Port: 443},
-				TLS: config.TLSConfig{
-					Enabled:  true,
-					Provider: "letsencrypt",
-					Domain:   "app.example.com",
-				},
-			},
-			wantCurlCmd: "curl -sfk https://localhost:443/_vibewarden/health 2>/dev/null || curl -sf http://localhost:443/_vibewarden/health",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			executor := &fakeExecutor{}
-			generator := &fakeGenerator{}
-
-			svc := deployapp.NewService(executor, generator)
-
-			var buf bytes.Buffer
-			err := svc.Deploy(context.Background(), tt.cfg, deployapp.RunOptions{
-				ConfigPath:   "/tmp/proj/vibewarden.yaml",
-				GeneratedDir: t.TempDir(),
-				Out:          &buf,
-			})
-			if err != nil {
-				t.Fatalf("Deploy() unexpected error: %v", err)
-			}
-
-			assertRunCalled(t, executor.runCalls, tt.wantCurlCmd)
-		})
-	}
-}
-
-// TestService_Deploy_RsyncSourceIsAlwaysBundleDir verifies that the rsync
-// source is always the deploy bundle directory, never CWD or parent-of-CWD.
-func TestService_Deploy_RsyncSourceIsAlwaysBundleDir(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	bundleDir := t.TempDir()
-
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/proj/vibewarden.yaml",
-		ProjectName:  "myproject",
-		GeneratedDir: bundleDir,
-	})
-	if err != nil {
-		t.Fatalf("Deploy() unexpected error: %v", err)
-	}
-
-	if len(executor.transferCalls) == 0 {
-		t.Fatal("expected at least one Transfer call")
-	}
-
-	// The first Transfer call must use the bundle directory as the source.
-	firstTransfer := executor.transferCalls[0]
-	if firstTransfer.localDir != bundleDir {
-		t.Errorf("Transfer localDir = %q, want %q (the bundle directory)", firstTransfer.localDir, bundleDir)
-	}
-}
-
-// TestService_Deploy_NoTransferFileForConfig verifies that Deploy no longer
-// calls TransferFile for the config -- it is now inside the bundle.
-func TestService_Deploy_NoTransferFileForConfig(t *testing.T) {
-	executor := &fakeExecutor{}
-	generator := &fakeGenerator{}
-
-	svc := deployapp.NewService(executor, generator)
-
-	err := svc.Deploy(context.Background(), defaultConfig(), deployapp.RunOptions{
-		ConfigPath:   "/tmp/myproject/vibewarden.yaml",
-		ProjectName:  "myproject",
-		GeneratedDir: t.TempDir(),
-	})
-	if err != nil {
-		t.Fatalf("Deploy() unexpected error: %v", err)
-	}
-
-	// TransferFile should NOT be called -- config is bundled locally.
-	if len(executor.transferFileCalls) != 0 {
-		t.Errorf("expected no TransferFile calls (config is bundled), got %d: %v",
-			len(executor.transferFileCalls), executor.transferFileCalls)
-	}
-}
