+++ internal/cli/cmd/root.go
@@ -48,7 +48,6 @@ Zero-to-secure in minutes.`,
 	root.AddCommand(NewBuildCmd())
 	root.AddCommand(NewRestartCmd())
 	root.AddCommand(NewMCPCmd())
-	root.AddCommand(NewDeployCmd())
 	root.AddCommand(NewBundleCmd())
 	root.AddCommand(NewTLSCmd())
